+++ CHANGELOG.md
@@ -8,6 +8,7 @@ For more information about changelogs, check
 
 ## 1.3.0 - unreleased
 
+* [ENHANCEMENT] Bootstrappify forms wrapped in `navbar`, setting their class to "navbar-form"
 * [BUGFIX] Match default placeholder with label content (therefore supporting localization)
 * [ENHANCEMENT] Add `:help` option to display a help block after most form fields
 * [FEATURE] Extend `button_to` to be wrapped in <li> when inside a `nav`, just like `link_to`